requirements.txt
# Packages used when running tests. Installed by tox into multiple
# virtualenvs.
#
# To add/update dependencies, update test-requirements.txt.in (not the
# generated test-requirements.txt) and run make update-requirements

protobuf==3.3.0
grpcio==1.4.0
grpcio-tools==1.4.0
pysha3==1.0.2
pylint==1.7.2
tox==2.8.2
six==1.10.0
